requirements.txt
# Core scraping libraries
scrapy==2.11.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# Database libraries
pymongo==4.6.0
psycopg2-binary==2.9.9
redis==5.0.1

# Data processing
pandas==2.1.4
numpy==1.24.3
nltk==3.8.1
spacy==3.7.2

# Media processing
Pillow==10.1.0
opencv-python==4.8.1.78
ffmpeg-python==0.2.0

# Web APIs
google-api-python-client==2.108.0
tweepy==4.14.0
youtube-dl==2021.12.17

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
celery==5.3.4
flask==3.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional for dorking and reverse engineering
fake-useragent==1.4.0
requests-html==0.10.0
undetected-chromedriver==3.5.4
pyppeteer==2.0.0
lxml==4.9.3
cloudscraper==1.2.71